Update thorpej_scsipi to -current as of a month ago

+++ sys/compat/common/kern_ipc_10.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: kern_ipc_10.c,v 1.9 1999/08/25 04:47:12 thorpej Exp $	*/
+/*	$NetBSD: kern_ipc_10.c,v 1.9.2.1 2000/11/20 18:08:07 bouyer Exp $	*/
 
 /*
  * Copyright (c) 1994 Adam Glass and Charles M. Hannum.  All rights reserved.
@@ -43,11 +43,6 @@
 #include <sys/syscallargs.h>
 
 #include <compat/common/compat_util.h>
-
-#include <vm/vm.h>
-#include <vm/vm_map.h>
-#include <vm/vm_map.h>
-#include <vm/vm_kern.h>
 
 #ifdef SYSVSEM
 int
@@ -103,7 +98,8 @@ compat_10_sys_semsys(p, v, retval)
 
 	case 2:						/* semop() */
 		SCARG(&semop_args, semid) = SCARG(uap, a2);
-		SCARG(&semop_args, sops) = (struct sembuf *)SCARG(uap, a3);
+		SCARG(&semop_args, sops) =
+		    (struct sembuf *)(u_long)SCARG(uap, a3);
 		SCARG(&semop_args, nsops) = SCARG(uap, a4);
 		return (sys_semop(p, &semop_args, retval));
 
@@ -152,18 +148,21 @@ compat_10_sys_shmsys(p, v, retval)
 	switch (SCARG(uap, which)) {
 	case 0:						/* shmat() */
 		SCARG(&shmat_args, shmid) = SCARG(uap, a2);
-		SCARG(&shmat_args, shmaddr) = (void *)SCARG(uap, a3);
+		SCARG(&shmat_args, shmaddr) =
+		    (void *)(u_long)SCARG(uap, a3);
 		SCARG(&shmat_args, shmflg) = SCARG(uap, a4);
 		return (sys_shmat(p, &shmat_args, retval));
 
 	case 1:						/* shmctl() */
 		SCARG(&shmctl_args, shmid) = SCARG(uap, a2);
 		SCARG(&shmctl_args, cmd) = SCARG(uap, a3);
-		SCARG(&shmctl_args, buf) = (struct shmid_ds14 *)SCARG(uap, a4);
+		SCARG(&shmctl_args, buf) =
+		    (struct shmid_ds14 *)(u_long)SCARG(uap, a4);
 		return (compat_14_sys_shmctl(p, &shmctl_args, retval));
 
 	case 2:						/* shmdt() */
-		SCARG(&shmdt_args, shmaddr) = (void *)SCARG(uap, a2);
+		SCARG(&shmdt_args, shmaddr) =
+		    (void *)(u_long)SCARG(uap, a2);
 		return (sys_shmdt(p, &shmdt_args, retval));
 
 	case 3:						/* shmget() */
@@ -221,7 +220,7 @@ compat_10_sys_msgsys(p, v, retval)
 		SCARG(&msgctl_args, msqid) = SCARG(uap, a2);
 		SCARG(&msgctl_args, cmd) = SCARG(uap, a3);
 		SCARG(&msgctl_args, buf) =
-		    (struct msqid_ds14 *)SCARG(uap, a4);
+		    (struct msqid_ds14 *)(u_long)SCARG(uap, a4);
 		return (compat_14_sys_msgctl(p, &msgctl_args, retval));
 
 	case 1:					/* msgget() */
@@ -231,14 +230,16 @@ compat_10_sys_msgsys(p, v, retval)
 
 	case 2:					/* msgsnd() */
 		SCARG(&msgsnd_args, msqid) = SCARG(uap, a2);
-		SCARG(&msgsnd_args, msgp) = (void *)SCARG(uap, a3);
+		SCARG(&msgsnd_args, msgp) =
+		    (void *)(u_long)SCARG(uap, a3);
 		SCARG(&msgsnd_args, msgsz) = SCARG(uap, a4);
 		SCARG(&msgsnd_args, msgflg) = SCARG(uap, a5);
 		return (sys_msgsnd(p, &msgsnd_args, retval));
 
 	case 3:					/* msgrcv() */
 		SCARG(&msgrcv_args, msqid) = SCARG(uap, a2);
-		SCARG(&msgrcv_args, msgp) = (void *)SCARG(uap, a3);
+		SCARG(&msgrcv_args, msgp) =
+		    (void *)(u_long)SCARG(uap, a3);
 		SCARG(&msgrcv_args, msgsz) = SCARG(uap, a4);
 		SCARG(&msgrcv_args, msgtyp) = SCARG(uap, a5);
 		SCARG(&msgrcv_args, msgflg) = SCARG(uap, a6);
